Adapted install shell script from https://github.com/tleeuwenburg/pyconau15/blob/master/install.sh
Confirmed currently works with ./install.sh on OSX from commandline with Anaconda installed. Need to push 'y' now and again to confirm the install steps. Could consolidate and quiet this, but it gives some feedback on progress. Should work on Linux also but might need to confirm with some folks.